fix: 재고 API 오류 및 디지털 트윈 클릭 시스템 개선 company 추가 노션의 데이터 테이블에 맞게 수정 백엔드 오류 해결
새로운 기능 추가
Company 엔티티 및 DataInitService로 기본 회사 데이터 자동 생성
ClickEventController/Service로 Unity 디지털 트윈 클릭 이벤트 처리
CORS 설정을 allowedOriginPatterns로 개선하여 개발 환경 지원

코드 정리
Factory3DTwin 컴포넌트 로그 정리 및 리팩토링
사용하지 않는 엔티티/서비스 클래스 정리 (ProductPosition, RobotStatus 등)
프론트엔드 패키지 관리 파일 추가

+++ dashboard_backend/src/main/java/com/u1mobis/dashboard_backend/service/UserService.java
@@ -19,7 +19,7 @@ public class UserService {
         if (userRepository.findByUserName(userDTO.getUsername()) != null) {
             return false; // 이미 존재
         }
-        User user = new User(null, userDTO.getUsername(), passwordEncoder.encode(userDTO.getPassword()));
+        User user = new User((Long) null, userDTO.getUsername(), passwordEncoder.encode(userDTO.getPassword()));
         userRepository.save(user);
         return true;
     }
